requirements.txt
asgiref==3.5.0
astroid==2.11.3
backports.zoneinfo==0.2.1
colorama==0.4.4
dill==0.3.4
Django==4.0.4
django-crispy-forms==1.14.0
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysqlclient==2.1.0
platformdirs==2.5.2
pylint==2.13.7
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
wrapt==1.14.0
asgiref==3.5.0
auth==0.5.3
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
django-filter==21.1
dnspython==2.2.1
eventlet==0.33.0
falcon==3.1.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
mongoengine==0.24.1
mysqlclient==2.1.0
protobuf==3.20.1
pymongo==4.1.1
pytube==12.1.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
Werkzeug==2.1.2
